chore: improve code structure
