+++ client/src/app/components/stardust/block/payload/TransactionPayload.tsx
@@ -1,24 +1,12 @@
-import {
-    AddressUnlockCondition,
-    CommonOutput,
-    ExpirationUnlockCondition,
-    GovernorAddressUnlockCondition,
-    ReferenceUnlock,
-    SignatureUnlock,
-    StateControllerAddressUnlockCondition,
-    UnlockConditionType,
-    UnlockType,
-    Utils,
-} from "@iota/sdk-wasm/web";
 import React, { useContext, useEffect, useState } from "react";
 import NetworkContext from "~/app/context/NetworkContext";
-import { Bech32AddressHelper } from "~/helpers/stardust/bech32AddressHelper";
 import transactionPayloadMessage from "~assets/modals/stardust/block/transaction-payload.json";
+import { getInputsPreExpandedConfig, getOutputsPreExpandedConfig } from "~helpers/stardust/preExpandedConfig";
+import { IPreExpandedConfig } from "~models/components";
 import Modal from "../../../Modal";
 import Input from "../../Input";
 import Output from "../../Output";
 import Unlocks from "../../Unlocks";
-import { IPreExpandedConfig } from "../../interfaces";
 import "./TransactionPayload.scss";
 import { TransactionPayloadProps } from "./TransactionPayloadProps";
 
@@ -29,103 +17,14 @@ const TransactionPayload: React.FC<TransactionPayloadProps> = ({ network, inputs
     const [inputsPreExpandedConfig, setInputsPreExpandedConfig] = useState<IPreExpandedConfig[]>([]);
     const { bech32Hrp } = useContext(NetworkContext);
 
-    const OUTPUT_EXPAND_CONDITIONS: UnlockConditionType[] = [
-        UnlockConditionType.Address,
-        UnlockConditionType.StateControllerAddress,
-        UnlockConditionType.GovernorAddress,
-    ];
-
-    const INPUT_EXPAND_CONDITIONS: UnlockConditionType[] = [...OUTPUT_EXPAND_CONDITIONS, UnlockConditionType.Expiration];
+    const outputsPreExpandedConfig = getOutputsPreExpandedConfig(outputs);
 
     useEffect(() => {
         if (bech32Hrp) {
-            // given the unlocks, expand the correct address unlock condition
-            const inputsPreExpandedConfig: IPreExpandedConfig[] = inputs.map((input, idx) => {
-                const commonOutput = input?.output?.output as unknown as CommonOutput;
-                let preExpandedConfig: IPreExpandedConfig = {};
-                if (commonOutput) {
-                    const matchExpandCondition = commonOutput.unlockConditions?.find((unlockCondition) =>
-                        INPUT_EXPAND_CONDITIONS.includes(unlockCondition.type),
-                    );
-                    preExpandedConfig = {
-                        isPreExpanded: !!matchExpandCondition,
-                    };
-                    if (input?.output?.output && "unlockConditions" in input.output.output) {
-                        const commmonOutput = input.output.output as unknown as CommonOutput;
-                        let unlock = unlocks[idx];
-                        if (unlock.type === UnlockType.Reference) {
-                            const referenceUnlock = unlock as ReferenceUnlock;
-                            unlock = unlocks[referenceUnlock.reference];
-                        }
-                        const unlockSignatureAddress = Utils.hexPublicKeyToBech32Address(
-                            (unlock as SignatureUnlock).signature.publicKey,
-                            bech32Hrp,
-                        );
-                        preExpandedConfig = {
-                            ...preExpandedConfig,
-                            unlockConditions: commmonOutput.unlockConditions?.map((unlockCondition) => {
-                                switch (unlockCondition.type) {
-                                    case UnlockConditionType.Address: {
-                                        const unlockAddress = Bech32AddressHelper.buildAddress(
-                                            bech32Hrp,
-                                            (unlockCondition as AddressUnlockCondition).address,
-                                        )?.bech32;
-                                        return unlockAddress === unlockSignatureAddress;
-                                    }
-                                    case UnlockConditionType.Expiration: {
-                                        const unlockAddress = Bech32AddressHelper.buildAddress(
-                                            bech32Hrp,
-                                            (unlockCondition as ExpirationUnlockCondition).returnAddress,
-                                        )?.bech32;
-                                        return unlockAddress === unlockSignatureAddress;
-                                    }
-                                    case UnlockConditionType.StateControllerAddress: {
-                                        const unlockAddress = Bech32AddressHelper.buildAddress(
-                                            bech32Hrp,
-                                            (unlockCondition as StateControllerAddressUnlockCondition).address,
-                                        )?.bech32;
-                                        return unlockAddress === unlockSignatureAddress;
-                                    }
-                                    case UnlockConditionType.GovernorAddress: {
-                                        const unlockAddress = Bech32AddressHelper.buildAddress(
-                                            bech32Hrp,
-                                            (unlockCondition as GovernorAddressUnlockCondition).address,
-                                        )?.bech32;
-                                        return unlockAddress === unlockSignatureAddress;
-                                    }
-                                    default:
-                                        return false;
-                                }
-                            }),
-                        };
-                    }
-                }
-                return preExpandedConfig;
-            });
+            const inputsPreExpandedConfig = getInputsPreExpandedConfig(inputs, unlocks, bech32Hrp);
             setInputsPreExpandedConfig(inputsPreExpandedConfig);
         }
     }, [bech32Hrp]);
-
-    // for basic outputs, always expand all the OUTPUT_EXPAND_CONDITIONS
-    const outputsPreExpandedConfig: IPreExpandedConfig[] = outputs.map((output) => {
-        const commonOutput = output.output as CommonOutput;
-        let preExpandedConfig: IPreExpandedConfig = {};
-        if (commonOutput) {
-            const matchExpandCondition = commonOutput.unlockConditions?.find((unlockCondition) =>
-                OUTPUT_EXPAND_CONDITIONS.includes(unlockCondition.type),
-            );
-            preExpandedConfig = {
-                isPreExpanded: !!matchExpandCondition,
-            };
-            preExpandedConfig = {
-                ...preExpandedConfig,
-                unlockConditions: commonOutput.unlockConditions?.map((unlockCondition) =>
-                    OUTPUT_EXPAND_CONDITIONS.includes(unlockCondition.type),
-                ),
-            };
-        }
-        return preExpandedConfig;
-    });
 
     return (
         <div className="transaction-payload">
