--- client/src/app/components/stardust/interfaces/index.tsx
@@ -1,1 +1,0 @@
-export * from './IPreExpandedConfig';
