+++ client/src/models/components/index.tsx
@@ -1,0 +1,1 @@
+export * from "./IPreExpandedConfig";
